requirements.txt
jinja2>=3.1.4
livereload>=2.7.0
pyyaml>=6.0.2
watchdog>=6.0.0
